Refactor and add doxygen comments
Reduce the duplication of code and make use of the ComplexMatrixBase
class.
Add doxygen comments to the start of each function.
